Add `jj sync` alias
Added `jj sync` as an alias for `jj git fetch` that will check if the
working copy is orphaned and if so move to a new revision from
'trunk()'. The use case is when I'm working on a bookmark and then it
gets merged in that revision can be left orphaned after doing `jj git
fetch`.